feat: move kube api to own package
